Move code to artifacts.rs
